requirements.txt
notion==0.0.25
gcsa==0.4.0
Flask==1.1.2
gunicorn==20.0.4
ipython==7.19.0
google-api-python-client==1.12.5
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
python-dotenv
SQLAlchemy==1.3.20
oauth2client==1.5.0
redis==3.5.3